All data records are blocks.
